Remove ProcessingTime entirely, move ProcessingTimeTrigger to Triggers.scala
